+++ accounts/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 # Register your models here.
-from accounts.models import ImitatedUser, User
+from accounts.models import Routine, User
 
-admin.site.register([User,ImitatedUser])
+admin.site.register([User, Routine])
