+++ accounts/models.py
@@ -6,8 +6,11 @@ class User(models.Model):
     def __str__(self):
         return self.uuid
 
-class ImitatedUser(models.Model):
-    user = models.ForeignKey('accounts.User', on_delete=models.CASCADE)
+class Routine(models.Model):
+    content = models.CharField(max_length=31)
+    is_completed = models.BooleanField(default=False)
+    imitated_user = models.ForeignKey('accounts.User', on_delete=models.CASCADE)
     celebrity = models.ForeignKey('celebrities.Celebrity', on_delete=models.CASCADE)
-    imitated_at = models.DateTimeField(auto_now_add=True)
 
+    def __str__(self):
+        return self.content
