+++ celebrities/models.py
@@ -1,3 +1,4 @@
+from django.contrib.postgres.fields import ArrayField
 from django.db import models
 
 class Celebrity(models.Model):
@@ -15,18 +16,7 @@ class Celebrity(models.Model):
     age = models.PositiveIntegerField()
     birthday = models.CharField(max_length=15)
     deceased_at = models.CharField(max_length=15)
-    imitated_users = models.ManyToManyField(
-        'accounts.User',
-        through='accounts.ImitateUser',
-        through_fields=('user', 'celebrity'), related_name='imitated_user')
-    routines = models.ManyToManyField('celebrities.Routine')
+    celebrity_routines = ArrayField(models.CharField(max_length=20), blank=True)
 
     def __str__(self):
         return self.name
-
-class Routine(models.Model):
-    content = models.CharField(max_length=31)
-    is_completed = models.BooleanField(default=False)
-
-    def __str__(self):
-        return self.content
